Adds a new option to signer app
With the option '-p' it is possible to sign the video with a
provisioned key. The key and certificate chain is read from
file and set through the corresponding Axis API.
If ONVIF Media Signing is installed, the video will be signed
with that for codecs H.26x.

This requires the library to have been installed with vendor
Axis.

Further, key, certificate chain and generator scripts have
been added to the test-files folder.